Clean up Devfs_Web into seperate files
