+++ packages/flutter_tools/test/general.shard/web/web_expression_compiler_test.dart
@@ -6,7 +6,7 @@ import 'package:dwds/dwds.dart';
 import 'package:file/memory.dart';
 import 'package:flutter_tools/src/base/file_system.dart';
 import 'package:flutter_tools/src/compile.dart';
-import 'package:flutter_tools/src/isolated/devfs_web.dart';
+import 'package:flutter_tools/src/isolated/web_expression_compiler.dart';
 import 'package:test/fake.dart';
 
 import '../../src/common.dart';
